Remove unused `import`.
